+++ src/integrationTest/java/uk/gov/hmcts/cmc/claimstore/aat/RestTestClient.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.cmc.claimstore.functional;
+package uk.gov.hmcts.cmc.claimstore.aat;
 
 import io.restassured.RestAssured;
 import io.restassured.response.Response;
