+++ src/integrationTest/java/uk/gov/hmcts/cmc/claimstore/aat/TestsSetup.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.cmc.claimstore.functional;
+package uk.gov.hmcts.cmc.claimstore.aat;
 
 public interface TestsSetup {
